seperated some actions into their own methods

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/subsystems/Arm.java
@@ -5,8 +5,8 @@ import com.qualcomm.robotcore.hardware.Servo;
 public class Arm {
     public static final String arm = "servoA";
 
-    public static final double ARM_UP = 1.0;
-    public static final double ARM_DOWN = 0.42;
+    public static final float ARM_UP = 1.0f;
+    public static final float ARM_DOWN = 0.42f;
 
     public Servo armServo = null;
 }
